requirements.txt
gunicorn==20.1.0
newspaper3k==0.2.8
Authlib==1.1.0
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==38.0.3
Flask==2.2.2
Flask-Authlib-Client==0.0.1
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.14.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.7.1
googleapis-common-protos==1.56.4
httplib2==0.21.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.1
msgpack==1.0.4
numpy
oauthlib==3.2.2
pandas
protobuf==4.21.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn
scipy
six
threadpoolctl
uritemplate
urllib3
Werkzeug
pyvi
beautifulsoup4
streamlit
pymongo
flask_pymongo
gevent
cloudinary
transformers
torch
sentencepiece